+++ src/traits.rs
@@ -5,7 +5,7 @@
 #[cfg(any(target_arch = "x86", target_arch = "x86_64", target_arch = "aarch64"))]
 use crate::enums::Reflector;
 
-use crate::structs::CrcParams;
+use crate::CrcParams;
 
 #[cfg(any(target_arch = "x86", target_arch = "x86_64", target_arch = "aarch64"))]
 use crate::structs::CrcState;